kubevirt: move vmware updaters to its own directory
